fw: instantiate a USB device
Once again, I built something that doesn't seem to work :(

As far as I can tell, not even the D+ pull-up is enabled, and the host
PC still doesn't know anything is connected to the port.

(edit 3 days later: seems to have been a layer 1 problem, where the side
of the MCU with D+ and D- pins didn't really get soldered down well.  A
bit of reflow and it's working fine now.)