fw: port to ATSAME51J-based board
Major differences from the SAM D21 code:

  * The D51-series doesn't have an 8MHz internal clock, instead, only a
    32K that the HAL configures to multiply up to 120MHz to run the CPU
    clock.

  * The oscillator configuration is in OSCCTRL now, not SYSCTRL.

  * The DPLLs must output at least 96MHz, so I've chosen to have them
    run at 96MHz and divide by two in the GCLK.

  * Microchip gave each GCLK its own register, instead of doing
    shenanigans with an "id" subfield; likewise a separate PCHCTRL for
    each peripheral output from the GCLK.

There are some interesting similarities between the implementations,
too, notably: most of the OSCCTRL (née SYSCTRL) and GCLK registers have
nearly the same fields.  Also of note is that the USB controller is
connected to the same GPIO pins across product families.

I looked at the implementation of with_internal_32kosc() to discover
that it configures DPLL0; my use of DPLL1 is dependent on this
implementation detail.